dfuse: remove needless sleeps
The DFU specification says that we should wait for bwPollTimeout after a
DFU_GETSTATUS if we are in dfuDNBUSY (in order to transition back to
dfuDNLOADSYNC) not in the other cases.
Skipping these waits saves us a lot of time when flashing a STM32
(about half to the total time) and flashing looks still reliable.