dfuse: [HACK] do not wait after setting address
The DFU specification says that we should wait for bwPollTimeout after a
DFU_GETSTATUS. On the specific ST USB bootloader protocol (aka dfuse)
the flash address is set by using the proprietary SET_ADDRESS command
transported using a DFU_DNLOAD packet. So we are the large delay on all
SET_ADDRESS, this is useless in most cases as the actual flashing is
not performed at this stage (unless we are sending incomplete chunks).
Skipping these waits saves us significant time but this is tweaking the
official DFU spec and corner cases with incomplete sectors might not
work.